@@ -1,14 +1,19 @@
-
 class Information {
-  var deaths = 0;
-  var recovered = 0;
-  var total = 0;
+  int deaths;
+  int recovered;
+  int total;
 
-  Information({this.total, this.recovered, this.deaths});
+  Information({
+    required this.total,
+    required this.recovered,
+    required this.deaths,
+  });
 
-  Information.fromJson(Map<String, dynamic> json) {
-    deaths = json["deaths"];
-    recovered = json["recovered"];
-    total = json["total"];
+  factory Information.fromJson(Map<String, dynamic> json) {
+    return Information(
+      deaths: json["deaths"],
+      recovered: json["recovered"],
+      total: json["total"],
+    );
   }
 }
